+++ src/geom/objects/mod.rs
@@ -1,0 +1,5 @@
+mod line;
+
+pub use self::{
+    line::Line,
+};